bot/requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
python-dotenv==1.0.1
requests==2.31.0
httpx==0.27.0

# Лёгкий RAG (без chroma!)

